refactor: extract compaction types and service to closeclaw-session crate
Move CompactionResult, TokenWarningState, CompactionError, CompactionService,
CompactionMessage, helper functions, and constants from src/session/compaction.rs
to crates/session/src/compaction.rs. The LLM-calling execute_compact function
remains in the main crate as a thin wrapper, bridging llm::Message to
CompactionMessage. CompactionError::LLMCallFailed now wraps a String instead
of LLMError to decouple from the llm crate.

Source: issue #1252
Type: refactor
