+++ crates/session/src/lib.rs
@@ -6,6 +6,7 @@
 
 pub mod bootstrap;
 pub mod checkpoint_manager;
+pub mod compaction;
 pub mod events;
 pub mod persistence;
 pub mod recovery;
